* fixes BUG ID: 3033335 * testing revised * other minor changes

+++ trunk/src/main/com/siemens/ct/exi/io/channel/BitDecoderChannel.java
@@ -56,7 +56,7 @@ public class BitDecoderChannel extends AbstractDecoderChannel implements
 		istream.align();
 	}
 	
-	public void skip(int n) throws IOException {
+	public void skip(long n) throws IOException {
 		istream.skip(n);
 	}
